refactor/remove chat feature
