+++ CONTRIBUTING.md
@@ -43,7 +43,6 @@ Common scopes for this project include:
 | **achievement-manager**  | Changes to the achievement manager components and hooks                                      |
 | **trading-card-manager** | Changes to the trading card manager components and hooks                                     |
 | **auto-idle**            | Changes to the auto-idle components and hooks                                                |
-| **chat**                 | Changes to the chat components and hooks                                                     |
 | **automation**           | Changes that affect multiple automation features (e.g.: card farming, achievement unlocker)  |
 | **settings**             | Changes to the settings components and hooks                                                 |
 | **states**               | Changes to nextjs frontend global contexts                                                   |
