+++ docs/content/changelog/2.5.13.mdx
@@ -1,0 +1,8 @@
+---
+title: 2.5.13
+date: 2025-12-04
+tags: ['Improved']
+---
+
+### Improved 🚀
+- Removed the "**Chatbox**" feature from the application due to ongoing issues. Users looking for community interaction are encouraged to join our official Discord server https://discord.com/invite/eB3uRXZ3Am.
