+++ docs/content/docs/settings/customization.mdx
@@ -10,9 +10,6 @@ Disable non-essential tooltips in the app
 
 Tooltips are small pop-up messages that pop up when you hover over certain elements in the app. They provide additional information about the element, but can be distracting for some users
 
-## Chat Notification Volume
-Set the volume for chat notification sounds when someone mentions you in a chat message
-
 ## Background Image
 Set a custom background image for the SGI main window. You can choose any image file from your computer to personalize the look of the app
