+++ src-tauri/src/settings.rs
@@ -18,7 +18,6 @@ fn get_default_settings() -> Value {
             "runAtStartup": false,
             "startMinimized": false,
             "closeToTray": true,
-            "chatSounds": [0.5],
             "customBackground": null,
             "autoRedeemFreeGames": false
         },
