+++ src/components/settings/Settings.tsx
@@ -13,9 +13,9 @@ import FreeGamesSettings from '@/components/settings/FreeGamesSettings'
 import GameSettings from '@/components/settings/GameSettings'
 import GeneralSettings from '@/components/settings/GeneralSettings'
 import Logs from '@/components/settings/Logs'
+import SocialButtons from '@/components/settings/SocialButtons'
 import SteamCredentials from '@/components/settings/SteamCredentials'
 import TradingCardManagerSettings from '@/components/settings/TradingCardManagerSettings'
-import ViewDocumentation from '@/components/settings/ViewDocumentation'
 import AdSlot from '@/components/ui/AdSlot'
 import useSettings from '@/hooks/settings/useSettings'
 
@@ -47,7 +47,7 @@ export default function Settings(): ReactElement {
       <div className='absolute flex flex-col items-center gap-4 bottom-4 left-0 px-6 w-[250px] z-40'>
         {process.env.NODE_ENV === 'production' && <AdSlot />}
 
-        <ViewDocumentation />
+        <SocialButtons />
         <span className='text-xs text-altwhite text-center'>Steam Game Idler v{version}</span>
       </div>
