+++ src/i18n/locales/en-US/translation.json
@@ -14,6 +14,7 @@
   "menu.issue": "Report an Issue",
   "menu.feature": "Request a Feature",
   "menu.support": "Support the Project",
+  "menu.joinDiscord": "Join our Discord",
   "menu.changelog": "View Changelog",
   "menu.update": "Check for Updates..",
   "sidebar.section.games": "Games",
@@ -133,12 +134,6 @@
   "achievementManager.statistics.saveChanges": "Save Changes",
   "achievementManager.statistics.resetAll": "Reset All",
   "achievementManager.statistics.flags": "Flags",
-  "chat.title": "Chatbox",
-  "chat.inputPlaceholder": "Send a message...",
-  "chat.addTextToImage": "Add a message (optional)...",
-  "chat.messageAction.delete": "delete this message",
-  "chat.messageAction.ban": "ban/unban this user",
-  "chat.messageAction.modal": "Are you sure you want to <1>{{action}}</1>?",
   "settings.title": "Settings",
   "settings.general.title": "General",
   "settings.general.displayName": "Display Name",
@@ -151,8 +146,6 @@
   "settings.general.startMinimized.description": "Start SGI minimized to the system tray",
   "settings.general.closeToTray": "Close to Tray",
   "settings.general.closeToTray.description": "Clicking the X button will minimize SGI to the system tray instead of closing it",
-  "settings.general.chatSounds": "Chat Notification Volume",
-  "settings.general.chatSounds.description": "Set the volume for chat notification sounds to {{value}}",
   "settings.general.useBeta": "Beta Features",
   "settings.general.useBeta.description": "Beta features may be unstable, change unexpectedly, or be removed at any time. If you experience any issues when using a beta feature, please report it on GitHub",
   "settings.general.disableTooltips": "Disable Tooltips",
@@ -237,7 +230,7 @@
   "proMode.modal.credentials": "Automated Steam credentials retrieval",
   "proMode.modal.autoRedeem": "Automatically redeem free games",
   "proMode.modal.themes": "Theme customization",
-  "proMode.modal.uniqueChatRole": "Unique chat role and color",
+  "proMode.modal.uniqueChatRole": "Discord role and color",
   "proMode.modal.moreBenefits": "More benefits coming in the future",
   "proMode.modal.bestValue": "Best Value",
   "proMode.modal.footer": "Prices displayed are in USD and exclude taxes.",
