+++ src/stores/userStore.ts
@@ -63,7 +63,6 @@ export const useUserStore = create<UserStore>(set => ({
       runAtStartup: false,
       startMinimized: false,
       closeToTray: true,
-      chatSounds: [0.5],
       customBackground: null,
       autoRedeemFreeGames: false,
     },
