+++ src/styles/globals.css
@@ -148,159 +148,6 @@ svg {
   }
 }
 
-/* Chat feature styles */
-
-.message-render::-webkit-scrollbar-track {
-  background: none;
-}
-
-.user-render::-webkit-scrollbar-track {
-  background: none;
-}
-
-.chat-message .markdown-body {
-  background-color: transparent !important;
-  background: inherit;
-  color: inherit !important;
-  font-size: inherit !important;
-  font-family: inherit !important;
-  line-height: inherit !important;
-}
-
-.chat-message .markdown-body pre code {
-  white-space: pre-wrap !important;
-  /* line-break: anywhere !important; */
-  overflow-wrap: break-word !important;
-}
-
-.chat-message a {
-  color: #51a2ff !important;
-}
-
-.chat-message a:hover {
-  color: #4592eaff !important;
-}
-
-.chat-message img:hover {
-  opacity: 0.9;
-}
-
-.chat-message p {
-  margin-top: 4px !important;
-  margin-bottom: 12px !important;
-  white-space: break-spaces;
-  word-wrap: break-word;
-}
-
-.chat-message strong {
-  font-weight: 800 !important;
-}
-
-.chat-message ul {
-  list-style-position: inside;
-  counter-reset: list-counter;
-  list-style: disc;
-}
-
-.chat-message ol {
-  list-style-position: inside;
-  counter-reset: list-counter;
-  list-style: decimal;
-}
-
-.chat-message li ul {
-  list-style-type: circle;
-}
-
-.chat-message li ul ul,
-.chat-message li ol ul {
-  list-style-type: square;
-}
-
-.chat-message blockquote {
-  border-left: 3px solid #474747ff !important;
-  color: #8a8a8aff !important;
-  font-style: italic !important;
-}
-
-.chat-message pre {
-  background-color: #232323ff !important;
-  padding: 0.5rem 0.4rem !important;
-  border: 1px solid #313131ff !important;
-  border-radius: 0.2rem !important;
-  font-size: 0.95em !important;
-  max-width: 70%;
-}
-
-.chat-message pre code {
-  font-family: 'Fira Code', monospace !important;
-  font-size: 0.7rem !important;
-  margin-bottom: 10px !important;
-}
-
-.chat-message code:not(pre code) {
-  background-color: #232323ff !important;
-  border: 1px solid #313131ff !important;
-  font-family: 'Fira Code', monospace !important;
-  font-size: 0.7rem !important;
-}
-
-.chat-message .mention-highlight {
-  background: rgba(77, 112, 252, 0.36);
-  color: #f1f8ffff;
-  padding: 2px;
-  border-radius: 2px;
-}
-
-.emoji-only-message {
-  font-size: 2rem;
-  margin-top: 12px;
-  margin-bottom: 12px;
-}
-
-.emoji:not(.emoji-only-message .emoji) {
-  font-size: 1.4em;
-  vertical-align: middle;
-  line-height: 1;
-  user-select: none;
-}
-
-.donator-role {
-  background: linear-gradient(90deg, #9645ffff, #3f4cffff, #00bfffff, #00bfffff, #3f4cffff, #9645ffff);
-  background-size: 200% auto;
-  background-clip: text;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  animation: gradient-shift 3.5s linear infinite;
-}
-
-.donator-badge {
-  display: inline-block;
-  -webkit-text-fill-color: #ffffff;
-}
-
-@keyframes gradient-shift {
-  0% {
-    background-position: 0% center;
-  }
-  100% {
-    background-position: 200% center;
-  }
-}
-
-.message-highlight {
-  animation: message-flash 2s ease-out;
-}
-
-@keyframes message-flash {
-  0% {
-    background-color: rgba(77, 112, 252, 0.4);
-  }
-  100% {
-    background-color: transparent;
-  }
-}
-
 /* Chatway */
 
 .chatway--frame-container {
