+++ src/utils/tasks.ts
@@ -194,40 +194,6 @@ export async function isPortableCheck(): Promise<boolean> {
   }
 }
 
-export async function playMentionBeep(): Promise<void> {
-  try {
-    const userSummary = JSON.parse(localStorage.getItem('userSummary') || '{}') as UserSummary
-
-    const response = await invoke<InvokeSettings>('get_user_settings', {
-      steamId: userSummary?.steamId,
-    })
-
-    const settings = response.settings
-
-    const { chatSounds } = settings?.general || {}
-
-    // Use Web Audio API for volume amplification
-    const audioContext = new AudioContext()
-    const audioResponse = await fetch('/pop.mp3')
-    const arrayBuffer = await audioResponse.arrayBuffer()
-    const audioBuffer = await audioContext.decodeAudioData(arrayBuffer)
-
-    const source = audioContext.createBufferSource()
-    source.buffer = audioBuffer
-
-    const gainNode = audioContext.createGain()
-    // Amplify volume - can go beyond 1.0 (e.g., 2.0 = 200%)
-    const volumeMultiplier = chatSounds[0]
-    gainNode.gain.value = volumeMultiplier
-
-    source.connect(gainNode)
-    gainNode.connect(audioContext.destination)
-    source.start(0)
-  } catch (error) {
-    console.error('Error playing mention beep:', error)
-  }
-}
-
 export const handleOpenExtLink = async (href: string): Promise<void> => {
   try {
     await open(href)
