+++ tailwind.config.mjs
@@ -65,7 +65,6 @@ const config = {
             'border': '#1f1f1fff',
             'input': '#333333ff',
             'inputhover': '#272727ff',
-            'chat-input': '#202020ff',
             'switch': '#333333ff',
             'tab-panel': '#00000058',
             'popover': '#101010ff',
@@ -104,7 +103,6 @@ const config = {
             'border': '#18202b',
             'input': '#1b232dff',
             'inputhover': '#182029ff',
-            'chat-input': '#121820ff',
             'switch': '#202934ff',
             'tab-panel': '#070a0f55',
             'popover': '#0c1016',
@@ -143,7 +141,6 @@ const config = {
             'border': '#241518',
             'input': '#23181aff',
             'inputhover': '#1d1315ff',
-            'chat-input': '#1a1113ff',
             'switch': '#2a1d20ff',
             'tab-panel': '#0c070955',
             'popover': '#10090b',
@@ -182,7 +179,6 @@ const config = {
             'border': '#221c2f',
             'input': '#1e1a27ff',
             'inputhover': '#1a1723ff',
-            'chat-input': '#16131eff',
             'switch': '#231f2eff',
             'tab-panel': '#0a091055',
             'popover': '#0f0d15',
@@ -221,7 +217,6 @@ const config = {
             'border': '#2b1b28',
             'input': '#2d1f2eff',
             'inputhover': '#281b29ff',
-            'chat-input': '#16131eff',
             'switch': '#2c1f2eff',
             'tab-panel': '#0a091055',
             'popover': '#0f0d15',
@@ -260,7 +255,6 @@ const config = {
             'border': '#2b2213',
             'input': '#3a3221ff',
             'inputhover': '#332c1dff',
-            'chat-input': '#121212ff',
             'switch': '#463d29ff',
             'tab-panel': '#05050555',
             'popover': '#0a0a0a',
@@ -299,7 +293,6 @@ const config = {
             'border': '#1a1a1a',
             'input': '#111111ff',
             'inputhover': '#0c0c0cff',
-            'chat-input': '#111111ff',
             'switch': '#161616ff',
             'tab-panel': '#00000088',
             'popover': '#050505',
